kern: uprobe: 优化

+++ Makefile
@@ -3,9 +3,14 @@ CWD = $(shell pwd)
 UNAME_A := $(shell uname -a)
 
 all: module tools java_agent pkg
+ifneq ($(findstring Ubuntu,$(UNAME_A)),)
+	dpkg -P diagnose-tools || echo "remove diagnose-tools error"
+	cd rpmbuild; sudo dpkg -i diagnose-tools*.deb
+else
 	yum remove -y diagnose-tools
 	yum install -y rpmbuild/RPMS/x86_64/diagnose-tools-*.rpm
 	diagnose-tools -v
+endif
 
 devel:
 ifneq ($(findstring Ubuntu,$(UNAME_A)),)
@@ -14,6 +19,9 @@ ifneq ($(findstring Ubuntu,$(UNAME_A)),)
 	apt -y install libunwind8-dev
 	apt -y install elfutils
 	apt -y install libelf-dev
+	apt -y install rpm
+	apt -y install alien
+	apt -y install bash-completion # git自动补全
 	apt install openjdk-8-jdk
 else
 	yum install -y libstdc++-static
@@ -54,17 +62,16 @@ java_agent:
 pkg:
 	cd rpmbuild; sh rpmbuild.sh
 	ls rpmbuild/RPMS/*/*
-
-deb:
-	rm ./rpmbuild/diagnose-tools*.deb
-	sudo alien -d ./rpmbuild/RPMS/x86_64/diagnose-tools*.rpm
-	#dpkg -P diagnose-tools || echo "remove alibaba diagnose tool error"
-	#sudo dpkg -i diagnose-tools*.deb
+ifneq ($(findstring Ubuntu,$(UNAME_A)),)
+	#sudo dpkg-reconfigure dash !!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	cd rpmbuild; rm -f diagnose-tools*.deb
+	cd rpmbuild; sudo alien -d ./RPMS/x86_64/diagnose-tools*.rpm
+endif
 
 test:
 	modprobe ext4
 	insmod SOURCE/module/diagnose.ko || echo ""
-	bash ./SOURCE/script/test.sh
+	bash ./SOURCE/script/test.sh $(case)
 	rmmod diagnose
 	rm tmp.txt -f
 	rm *.svg -f
